refactor: changes to run app in docker container

+++ src/main/resources/db/migration/V2024.03.23.2250__create_courses_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE courses
     created_at    TIMESTAMP    NOT NULL DEFAULT NOW(),
     name          VARCHAR(100) NOT NULL,
     code          VARCHAR(10)  NOT NULL,
-    instructor_id BIGINT       NOT NULL,
+    instructor_id INT       NOT NULL,
     description   TEXT         NOT NULL,
     status ENUM('ACTIVE', 'INACTIVE') NOT NULL DEFAULT 'ACTIVE',
     inactivated_at TIMESTAMP,
