+++ src/main/resources/db/migration/V2024.03.24.1120__create_registrations_table.sql
@@ -2,8 +2,8 @@ CREATE TABLE registrations
 (
     id            INT    PRIMARY KEY AUTO_INCREMENT,
     created_at    TIMESTAMP    NOT NULL DEFAULT NOW(),
-    user_id       BIGINT NOT NULL,
-    course_id     BIGINT  NOT NULL,
+    user_id       INT NOT NULL,
+    course_id     INT  NOT NULL,
 
     CONSTRAINT registration_user_id_course_id_ukey UNIQUE (user_id, course_id),
     CONSTRAINT registration_user_id_fkey FOREIGN KEY (user_id) REFERENCES users (id),
