+++ src/main/resources/db/migration/V2024.03.24.1530__create_course_reviews_table.sql
@@ -2,8 +2,8 @@ CREATE TABLE course_reviews
 (
     id            INT    PRIMARY KEY AUTO_INCREMENT,
     created_at    TIMESTAMP  NOT NULL DEFAULT NOW(),
-    user_id       BIGINT     NOT NULL,
-    course_id     BIGINT     NOT NULL,
+    user_id       INT     NOT NULL,
+    course_id     INT     NOT NULL,
     rating        INTEGER    NOT NULL,
     comment       TEXT,
